Add display formatting for number renderer
When the px-data-grid-number-renderer is used for a column the
developer can pass new rendererConfig options, including
rendererConfig.displayFormat, which should be set to a valid
numbro.js string used to format the number before it is displayed.

Additional renderer options are exposed to customize how numbro.js
formats the value. See the px-data-grid-renderer-number.html file
for a full list.

Additional changes:

* Uses parseFloat instead of parseInt to parse numbers, which
  supports user input data formatted as a string containing
  floats or integers.
* Throws an error when given a value that cannot be parsed
  to a number, but recovers and renders the value anyway.
* Uses an unreleased version of px-number-formatter, will
  remove this dependency and update to a real semver tag once
  the number formatter changes are merged